Divide el service de Factura. Mueve models DTO a su package correspondiente.
