+++ src/main/java/sic/controller/PedidoController.java
@@ -17,8 +17,8 @@ import sic.aspect.AccesoRolesPermitidos;
 import sic.exception.BusinessServiceException;
 import sic.modelo.*;
 import sic.modelo.criteria.BusquedaPedidoCriteria;
-import sic.modelo.calculos.NuevosResultadosComprobanteDTO;
-import sic.modelo.calculos.Resultados;
+import sic.modelo.dto.NuevosResultadosComprobanteDTO;
+import sic.modelo.Resultados;
 import sic.modelo.dto.PedidoDTO;
 import sic.modelo.dto.NuevoRenglonPedidoDTO;
 import sic.service.*;
