+++ src/main/java/sic/modelo/dto/NuevaFacturaVentaDTO.java
@@ -1,9 +1,10 @@
-package sic.modelo;
+package sic.modelo.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import sic.modelo.TipoDeComprobante;
 import sic.modelo.dto.NuevoRenglonFacturaDTO;
 
 import java.math.BigDecimal;
