+++ src/main/java/sic/service/impl/FacturaServiceImpl.java
@@ -1,72 +1,41 @@
 package sic.service.impl;
 
-import java.io.IOException;
-
-import com.querydsl.core.BooleanBuilder;
 import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.validation.annotation.Validated;
 import sic.modelo.*;
-
-import java.io.InputStream;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
-import java.net.URL;
 import java.time.LocalDateTime;
 import java.util.*;
-import javax.imageio.ImageIO;
 import javax.persistence.EntityNotFoundException;
-import javax.swing.ImageIcon;
-import javax.validation.Valid;
-
-import net.sf.jasperreports.engine.JRException;
-import net.sf.jasperreports.engine.JasperExportManager;
-import net.sf.jasperreports.engine.JasperFillManager;
-import net.sf.jasperreports.engine.data.JRBeanCollectionDataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sic.modelo.calculos.NuevosResultadosComprobanteDTO;
-import sic.modelo.calculos.Resultados;
-import sic.modelo.criteria.BusquedaFacturaCompraCriteria;
-import sic.modelo.criteria.BusquedaFacturaVentaCriteria;
+import sic.modelo.dto.NuevosResultadosComprobanteDTO;
+import sic.modelo.Resultados;
 import sic.service.*;
 import sic.exception.BusinessServiceException;
-import sic.exception.ServiceException;
-import sic.util.CalculosComprobante;
-import sic.repository.FacturaVentaRepository;
-import sic.repository.FacturaCompraRepository;
 import sic.repository.FacturaRepository;
 import sic.repository.RenglonFacturaRepository;
+import sic.util.CalculosComprobante;
 
 @Service
 @Validated
 public class FacturaServiceImpl implements IFacturaService {
 
   private final FacturaRepository facturaRepository;
-  private final FacturaVentaRepository facturaVentaRepository;
-  private final FacturaCompraRepository facturaCompraRepository;
   private final RenglonFacturaRepository renglonFacturaRepository;
   private final IProductoService productoService;
-  private final IConfiguracionSucursalService configuracionSucursalService;
   private final IPedidoService pedidoService;
   private final INotaService notaService;
   private final ICuentaCorrienteService cuentaCorrienteService;
-  private final IAfipService afipService;
-  private final IReciboService reciboService;
-  private final IUsuarioService usuarioService;
-  private final IClienteService clienteService;
-  private final ICorreoElectronicoService correoElectronicoService;
   private static final BigDecimal IVA_21 = new BigDecimal("21");
   private static final BigDecimal IVA_105 = new BigDecimal("10.5");
   private static final BigDecimal CIEN = new BigDecimal("100");
-  private final Logger logger = LoggerFactory.getLogger(this.getClass());
   private static final int TAMANIO_PAGINA_DEFAULT = 25;
   private final MessageSource messageSource;
 
@@ -74,34 +43,18 @@ public class FacturaServiceImpl implements IFacturaService {
   @Lazy
   public FacturaServiceImpl(
       FacturaRepository facturaRepository,
-      FacturaVentaRepository facturaVentaRepository,
-      FacturaCompraRepository facturaCompraRepository,
       RenglonFacturaRepository renglonFacturaRepository,
       IProductoService productoService,
-      IConfiguracionSucursalService configuracionSucursalService,
       IPedidoService pedidoService,
       INotaService notaService,
       ICuentaCorrienteService cuentaCorrienteService,
-      IAfipService afipService,
-      IReciboService reciboService,
-      IUsuarioService usuarioService,
-      IClienteService clienteService,
-      ICorreoElectronicoService correoElectronicoService,
       MessageSource messageSource) {
     this.facturaRepository = facturaRepository;
-    this.facturaVentaRepository = facturaVentaRepository;
-    this.facturaCompraRepository = facturaCompraRepository;
     this.renglonFacturaRepository = renglonFacturaRepository;
     this.productoService = productoService;
-    this.configuracionSucursalService = configuracionSucursalService;
     this.pedidoService = pedidoService;
     this.notaService = notaService;
     this.cuentaCorrienteService = cuentaCorrienteService;
-    this.afipService = afipService;
-    this.reciboService = reciboService;
-    this.usuarioService = usuarioService;
-    this.clienteService = clienteService;
-    this.correoElectronicoService = correoElectronicoService;
     this.messageSource = messageSource;
   }
 
@@ -151,71 +104,6 @@ public class FacturaServiceImpl implements IFacturaService {
   }
 
   @Override
-  public List<Factura> getFacturasDelPedido(Long idPedido) {
-    return facturaVentaRepository.findAllByPedidoAndEliminada(
-        pedidoService.getPedidoNoEliminadoPorId(idPedido), false);
-  }
-
-  @Override
-  public TipoDeComprobante[] getTipoFacturaCompra(Sucursal sucursal, Proveedor proveedor) {
-    if (CategoriaIVA.discriminaIVA(sucursal.getCategoriaIVA())) {
-      if (CategoriaIVA.discriminaIVA(proveedor.getCategoriaIVA())) {
-        TipoDeComprobante[] tiposPermitidos = new TipoDeComprobante[4];
-        tiposPermitidos[0] = TipoDeComprobante.FACTURA_A;
-        tiposPermitidos[1] = TipoDeComprobante.FACTURA_B;
-        tiposPermitidos[2] = TipoDeComprobante.FACTURA_X;
-        tiposPermitidos[3] = TipoDeComprobante.PRESUPUESTO;
-        return tiposPermitidos;
-      } else {
-        TipoDeComprobante[] tiposPermitidos = new TipoDeComprobante[3];
-        tiposPermitidos[0] = TipoDeComprobante.FACTURA_C;
-        tiposPermitidos[1] = TipoDeComprobante.FACTURA_X;
-        tiposPermitidos[2] = TipoDeComprobante.PRESUPUESTO;
-        return tiposPermitidos;
-      }
-    } else {
-      if (CategoriaIVA.discriminaIVA(proveedor.getCategoriaIVA())) {
-        TipoDeComprobante[] tiposPermitidos = new TipoDeComprobante[3];
-        tiposPermitidos[0] = TipoDeComprobante.FACTURA_B;
-        tiposPermitidos[1] = TipoDeComprobante.FACTURA_X;
-        tiposPermitidos[2] = TipoDeComprobante.PRESUPUESTO;
-        return tiposPermitidos;
-      } else {
-        TipoDeComprobante[] tiposPermitidos = new TipoDeComprobante[3];
-        tiposPermitidos[0] = TipoDeComprobante.FACTURA_C;
-        tiposPermitidos[1] = TipoDeComprobante.FACTURA_X;
-        tiposPermitidos[2] = TipoDeComprobante.PRESUPUESTO;
-        return tiposPermitidos;
-      }
-    }
-  }
-
-  @Override
-  public TipoDeComprobante[] getTipoFacturaVenta(Sucursal sucursal, Cliente cliente) {
-    if (CategoriaIVA.discriminaIVA(sucursal.getCategoriaIVA())) {
-      if (CategoriaIVA.discriminaIVA(cliente.getCategoriaIVA())) {
-        TipoDeComprobante[] tiposPermitidos = new TipoDeComprobante[3];
-        tiposPermitidos[0] = TipoDeComprobante.FACTURA_A;
-        tiposPermitidos[1] = TipoDeComprobante.FACTURA_X;
-        tiposPermitidos[2] = TipoDeComprobante.PRESUPUESTO;
-        return tiposPermitidos;
-      } else {
-        TipoDeComprobante[] tiposPermitidos = new TipoDeComprobante[3];
-        tiposPermitidos[0] = TipoDeComprobante.FACTURA_B;
-        tiposPermitidos[1] = TipoDeComprobante.FACTURA_X;
-        tiposPermitidos[2] = TipoDeComprobante.PRESUPUESTO;
-        return tiposPermitidos;
-      }
-    } else {
-      TipoDeComprobante[] tiposPermitidos = new TipoDeComprobante[3];
-      tiposPermitidos[0] = TipoDeComprobante.FACTURA_C;
-      tiposPermitidos[1] = TipoDeComprobante.FACTURA_X;
-      tiposPermitidos[2] = TipoDeComprobante.PRESUPUESTO;
-      return tiposPermitidos;
-    }
-  }
-
-  @Override
   public TipoDeComprobante[] getTiposFacturaSegunSucursal(Sucursal sucursal) {
     if (CategoriaIVA.discriminaIVA(sucursal.getCategoriaIVA())) {
       TipoDeComprobante[] tiposPermitidos = new TipoDeComprobante[5];
@@ -250,7 +138,8 @@ public class FacturaServiceImpl implements IFacturaService {
     return renglonFacturaRepository.findById(idRenglonFactura).orElse(null); // orElseThrow
   }
 
-  private Pageable getPageable(Integer pagina, String ordenarPor, String sentido) {
+  @Override
+  public Pageable getPageable(Integer pagina, String ordenarPor, String sentido) {
     if (pagina == null) pagina = 0;
     String ordenDefault = "fecha";
     if (ordenarPor == null || sentido == null) {
@@ -272,156 +161,7 @@ public class FacturaServiceImpl implements IFacturaService {
   }
 
   @Override
-  public Page<FacturaCompra> buscarFacturaCompra(BusquedaFacturaCompraCriteria criteria) {
-    return facturaCompraRepository.findAll(
-        this.getBuilderCompra(criteria),
-        this.getPageable(
-            (criteria.getPagina() == null || criteria.getPagina() < 0) ? 0 : criteria.getPagina(),
-            criteria.getOrdenarPor(),
-            criteria.getSentido()));
-  }
-
-  @Override
-  public Page<FacturaVenta> buscarFacturaVenta(
-      BusquedaFacturaVentaCriteria criteria, long idUsuarioLoggedIn) {
-    return facturaVentaRepository.findAll(
-        this.getBuilderVenta(criteria, idUsuarioLoggedIn),
-        this.getPageable(
-            (criteria.getPagina() == null || criteria.getPagina() < 0) ? 0 : criteria.getPagina(),
-            criteria.getOrdenarPor(),
-            criteria.getSentido()));
-  }
-
-  private BooleanBuilder getBuilderCompra(BusquedaFacturaCompraCriteria criteria) {
-    QFacturaCompra qFacturaCompra = QFacturaCompra.facturaCompra;
-    BooleanBuilder builder = new BooleanBuilder();
-    if (criteria.getIdSucursal() == null) {
-      throw new BusinessServiceException(
-        messageSource.getMessage(
-          "mensaje_busqueda_sin_sucursal", null, Locale.getDefault()));
-    }
-    builder.and(
-        qFacturaCompra
-            .sucursal
-            .idSucursal
-            .eq(criteria.getIdSucursal())
-            .and(qFacturaCompra.eliminada.eq(false)));
-    if (criteria.getFechaDesde() != null || criteria.getFechaHasta() != null) {
-      if (criteria.getFechaDesde() != null && criteria.getFechaHasta() != null) {
-        criteria.setFechaDesde(criteria.getFechaDesde().withHour(0).withMinute(0).withSecond(0));
-        criteria.setFechaHasta(criteria.getFechaHasta().withHour(23).withMinute(59).withSecond(59).withNano(999999999));
-        builder.and(
-          qFacturaCompra.fecha.between(criteria.getFechaDesde(), criteria.getFechaHasta()));
-      } else if (criteria.getFechaDesde() != null) {
-        criteria.setFechaDesde(criteria.getFechaDesde().withHour(0).withMinute(0).withSecond(0));
-        builder.and(qFacturaCompra.fecha.after(criteria.getFechaDesde()));
-      } else if (criteria.getFechaHasta() != null) {
-        criteria.setFechaHasta(criteria.getFechaHasta().withHour(23).withMinute(59).withSecond(59).withNano(999999999));
-        builder.and(qFacturaCompra.fecha.before(criteria.getFechaHasta()));
-      }
-    }
-    if (criteria.getIdProveedor() != null)
-      builder.and(qFacturaCompra.proveedor.idProveedor.eq(criteria.getIdProveedor()));
-    if (criteria.getTipoComprobante() != null)
-      builder.and(qFacturaCompra.tipoComprobante.eq(criteria.getTipoComprobante()));
-    if (criteria.getIdProducto() != null)
-      builder.and(qFacturaCompra.renglones.any().idProductoItem.eq(criteria.getIdProducto()));
-    if (criteria.getNumSerie() != null && criteria.getNumFactura() != null)
-      builder
-          .and(qFacturaCompra.numSerie.eq(criteria.getNumSerie()))
-          .and(qFacturaCompra.numFactura.eq(criteria.getNumFactura()));
-    return builder;
-  }
-
-  private BooleanBuilder getBuilderVenta(
-      BusquedaFacturaVentaCriteria criteria, long idUsuarioLoggedIn) {
-    QFacturaVenta qFacturaVenta = QFacturaVenta.facturaVenta;
-    BooleanBuilder builder = new BooleanBuilder();
-    if (criteria.getIdSucursal() == null) {
-      throw new BusinessServiceException(
-        messageSource.getMessage(
-          "mensaje_busqueda_sin_sucursal", null, Locale.getDefault()));
-    }
-    builder.and(
-        qFacturaVenta
-            .sucursal
-            .idSucursal
-            .eq(criteria.getIdSucursal())
-            .and(qFacturaVenta.eliminada.eq(false)));
-    if (criteria.getFechaDesde() != null || criteria.getFechaHasta() != null) {
-      if (criteria.getFechaDesde() != null && criteria.getFechaHasta() != null) {
-        criteria.setFechaDesde(criteria.getFechaDesde().withHour(0).withMinute(0).withSecond(0));
-        criteria.setFechaHasta(criteria.getFechaHasta().withHour(23).withMinute(59).withSecond(59).withNano(999999999));
-        builder.and(
-            qFacturaVenta.fecha.between(criteria.getFechaDesde(), criteria.getFechaHasta()));
-      } else if (criteria.getFechaDesde() != null) {
-        criteria.setFechaDesde(criteria.getFechaDesde().withHour(0).withMinute(0).withSecond(0));
-        builder.and(qFacturaVenta.fecha.after(criteria.getFechaDesde()));
-      } else if (criteria.getFechaHasta() != null) {
-        criteria.setFechaHasta(criteria.getFechaHasta().withHour(23).withMinute(59).withSecond(59).withNano(999999999));
-        builder.and(qFacturaVenta.fecha.before(criteria.getFechaHasta()));
-      }
-    }
-    if (criteria.getIdCliente() != null)
-      builder.and(qFacturaVenta.cliente.idCliente.eq(criteria.getIdCliente()));
-    if (criteria.getTipoComprobante() != null)
-      builder.and(qFacturaVenta.tipoComprobante.eq(criteria.getTipoComprobante()));
-    if (criteria.getIdUsuario() != null)
-      builder.and(qFacturaVenta.usuario.idUsuario.eq(criteria.getIdUsuario()));
-    if (criteria.getIdViajante() != null)
-      builder.and(qFacturaVenta.cliente.viajante.idUsuario.eq(criteria.getIdViajante()));
-    if (criteria.getNumSerie() != null && criteria.getNumFactura() != null)
-      builder
-          .and(qFacturaVenta.numSerie.eq(criteria.getNumSerie()))
-          .and(qFacturaVenta.numFactura.eq(criteria.getNumFactura()));
-    if (criteria.getNroPedido() != null)
-      builder.and(qFacturaVenta.pedido.nroPedido.eq(criteria.getNroPedido()));
-    if (criteria.getIdProducto() != null)
-      builder.and(qFacturaVenta.renglones.any().idProductoItem.eq(criteria.getIdProducto()));
-    Usuario usuarioLogueado = usuarioService.getUsuarioNoEliminadoPorId(idUsuarioLoggedIn);
-    BooleanBuilder rsPredicate = new BooleanBuilder();
-    if (!usuarioLogueado.getRoles().contains(Rol.ADMINISTRADOR)
-        && !usuarioLogueado.getRoles().contains(Rol.VENDEDOR)
-        && !usuarioLogueado.getRoles().contains(Rol.ENCARGADO)) {
-      usuarioLogueado
-          .getRoles()
-          .forEach(
-              rol -> {
-                if (rol == Rol.VIAJANTE) {
-                  rsPredicate.or(
-                      qFacturaVenta.cliente.viajante.idUsuario.eq(
-                          usuarioLogueado.getIdUsuario()));
-                }
-                if (rol == Rol.COMPRADOR) {
-                  Cliente clienteRelacionado =
-                      clienteService.getClientePorIdUsuario(idUsuarioLoggedIn);
-                  if (clienteRelacionado != null) {
-                    rsPredicate.or(
-                        qFacturaVenta.cliente.idCliente.eq(clienteRelacionado.getIdCliente()));
-                  } else {
-                    rsPredicate.or(qFacturaVenta.cliente.isNull());
-                  }
-                }
-              });
-      builder.and(rsPredicate);
-    }
-    return builder;
-  }
-
-  private Factura procesarFactura(Factura factura) {
-    factura.setEliminada(false);
-    if (factura instanceof FacturaVenta) {
-      factura.setFecha(LocalDateTime.now());
-      factura.setNumSerie(
-          configuracionSucursalService
-              .getConfiguracionSucursal(factura.getSucursal())
-              .getNroPuntoDeVentaAfip());
-      factura.setNumFactura(
-          this.calcularNumeroFacturaVenta(
-              factura.getTipoComprobante(),
-              factura.getNumSerie(),
-              factura.getSucursal().getIdSucursal()));
-    }
+  public Factura procesarFactura(Factura factura) {
     this.calcularCantidadDeArticulos(factura);
     this.validarOperacion(factura);
     return factura;
@@ -436,90 +176,22 @@ public class FacturaServiceImpl implements IFacturaService {
   }
 
   @Override
-  @Transactional
-  public List<FacturaVenta> guardar(
-      @Valid List<FacturaVenta> facturas, Long idPedido, List<Recibo> recibos) {
-    this.calcularValoresFacturasVentaAndActualizarStock(facturas);
-    List<FacturaVenta> facturasProcesadas = new ArrayList<>();
-    if (idPedido != null) {
-      Pedido pedido = pedidoService.getPedidoNoEliminadoPorId(idPedido);
-      facturas.forEach(f -> f.setPedido(pedido));
-      for (Factura f : facturas) {
-        FacturaVenta facturaGuardada =
-            facturaVentaRepository.save((FacturaVenta) this.procesarFactura(f));
-        this.cuentaCorrienteService.asentarEnCuentaCorriente(facturaGuardada, TipoDeOperacion.ALTA);
-        facturasProcesadas.add(facturaGuardada);
-        if (recibos != null) {
-          recibos.forEach(reciboService::guardar);
-        }
-      }
-      List<Factura> facturasParaRelacionarAlPedido = new ArrayList<>(facturasProcesadas);
-      pedidoService.actualizarFacturasDelPedido(pedido, facturasParaRelacionarAlPedido);
-      facturasProcesadas.forEach(f -> logger.warn("La Factura {} se guardó correctamente.", f));
-      pedidoService.actualizarEstadoPedido(pedido);
-    } else {
-      facturasProcesadas = new ArrayList<>();
-      for (Factura f : facturas) {
-        FacturaVenta facturaGuardada;
-        facturaGuardada = facturaVentaRepository.save((FacturaVenta) this.procesarFactura(f));
-        this.cuentaCorrienteService.asentarEnCuentaCorriente(facturaGuardada, TipoDeOperacion.ALTA);
-        facturasProcesadas.add(facturaGuardada);
-        logger.warn("La Factura {} se guardó correctamente.", facturaGuardada);
-        if (recibos != null) {
-          recibos.forEach(reciboService::guardar);
-        }
-      }
-    }
-    return facturasProcesadas;
-  }
-
-  @Override
-  @Transactional
-  public List<FacturaCompra> guardar(@Valid List<FacturaCompra> facturas) {
-    this.calcularValoresFacturasCompraAndActualizarStock(facturas);
-    List<FacturaCompra> facturasProcesadas = new ArrayList<>();
-    for (Factura f : facturas) {
-      FacturaCompra facturaGuardada = null;
-      if (f instanceof FacturaCompra) {
-        facturaGuardada = facturaCompraRepository.save((FacturaCompra) this.procesarFactura(f));
-        this.cuentaCorrienteService.asentarEnCuentaCorriente(facturaGuardada);
-      }
-      facturasProcesadas.add(facturaGuardada);
-    }
-    return facturasProcesadas;
-  }
-
-  private void calcularValoresFacturasVentaAndActualizarStock(List<FacturaVenta> facturas) {
-    facturas.forEach(
-        facturaVenta -> {
-          this.calcularValoresFactura(facturaVenta);
-          this.actualizarStock(facturaVenta, Movimiento.VENTA);
-        });
-  }
-
-  private void calcularValoresFacturasCompraAndActualizarStock(List<FacturaCompra> facturas) {
-    facturas.forEach(
-        facturaCompra -> {
-          this.calcularValoresFactura(facturaCompra);
-          this.actualizarStock(facturaCompra, Movimiento.COMPRA);
-        });
-  }
-
-  private void calcularValoresFactura(Factura factura) {
+  public void calcularValoresFactura(Factura factura) {
     BigDecimal[] importes = new BigDecimal[factura.getRenglones().size()];
     int i = 0;
     for (RenglonFactura renglon : factura.getRenglones()) {
       importes[i] = renglon.getImporte();
       i++;
     }
-    NuevosResultadosComprobanteDTO nuevosResultadosComprobante = new NuevosResultadosComprobanteDTO();
+    NuevosResultadosComprobanteDTO nuevosResultadosComprobante =
+        new NuevosResultadosComprobanteDTO();
     nuevosResultadosComprobante.setImporte(importes);
     nuevosResultadosComprobante.setDescuentoPorcentaje(factura.getDescuentoPorcentaje());
     nuevosResultadosComprobante.setRecargoPorcentaje(factura.getRecargoPorcentaje());
     i = 0;
     if (factura.getTipoComprobante() == TipoDeComprobante.FACTURA_A
         || factura.getTipoComprobante() == TipoDeComprobante.FACTURA_B
-    //    || factura.getTipoComprobante() == TipoDeComprobante.FACTURA_Y
+        //    || factura.getTipoComprobante() == TipoDeComprobante.FACTURA_Y
         || factura.getTipoComprobante() == TipoDeComprobante.PRESUPUESTO) {
       BigDecimal[] ivaPorcentajes = new BigDecimal[factura.getRenglones().size()];
       BigDecimal[] ivaNetos = new BigDecimal[factura.getRenglones().size()];
@@ -547,13 +219,14 @@ public class FacturaServiceImpl implements IFacturaService {
     factura.setTotal(resultadosFactura.getTotal());
   }
 
-  private void actualizarStock(Factura factura, Movimiento movimiento) {
+  @Override
+  public void actualizarStock(Factura factura, Movimiento movimiento) {
     productoService.actualizarStock(
-            this.getIdsProductosYCantidades(factura),
-            factura.getIdSucursal(),
-            TipoDeOperacion.ALTA,
-            movimiento,
-            factura.getTipoComprobante());
+        this.getIdsProductosYCantidades(factura),
+        factura.getIdSucursal(),
+        TipoDeOperacion.ALTA,
+        movimiento,
+        factura.getTipoComprobante());
   }
 
   private Map<Long, BigDecimal> getIdsProductosYCantidades(Factura factura) {
@@ -579,77 +252,6 @@ public class FacturaServiceImpl implements IFacturaService {
   }
 
   @Override
-  @Transactional
-  public FacturaVenta autorizarFacturaVenta(FacturaVenta fv) {
-    ComprobanteAFIP comprobante =
-        ComprobanteAFIP.builder()
-            .idComprobante(fv.getIdFactura())
-            .fecha(fv.getFecha())
-            .tipoComprobante(fv.getTipoComprobante())
-            .cae(fv.getCae())
-            .vencimientoCAE(fv.getVencimientoCae())
-            .numSerieAfip(fv.getNumSerieAfip())
-            .numFacturaAfip(fv.getNumFacturaAfip())
-            .sucursal(fv.getSucursal())
-            .cliente(fv.getClienteEmbedded())
-            .subtotalBruto(fv.getSubTotalBruto())
-            .iva105neto(fv.getIva105Neto())
-            .iva21neto(fv.getIva21Neto())
-            .montoNoGravado(BigDecimal.ZERO)
-            .total(fv.getTotal())
-            .build();
-    afipService.autorizar(comprobante);
-    fv.setCae(comprobante.getCae());
-    fv.setVencimientoCae(comprobante.getVencimientoCAE());
-    fv.setNumSerieAfip(comprobante.getNumSerieAfip());
-    fv.setNumFacturaAfip(comprobante.getNumFacturaAfip());
-    return fv;
-  }
-
-  @Override
-  public BigDecimal calcularTotalFacturadoVenta(
-      BusquedaFacturaVentaCriteria criteria, long idUsuarioLoggedIn) {
-    BigDecimal totalFacturado =
-        facturaVentaRepository.calcularTotalFacturadoVenta(
-            this.getBuilderVenta(criteria, idUsuarioLoggedIn));
-    return (totalFacturado != null ? totalFacturado : BigDecimal.ZERO);
-  }
-
-  @Override
-  public BigDecimal calcularTotalFacturadoCompra(BusquedaFacturaCompraCriteria criteria) {
-    BigDecimal totalFacturado =
-        facturaCompraRepository.calcularTotalFacturadoCompra(this.getBuilderCompra(criteria));
-    return (totalFacturado != null ? totalFacturado : BigDecimal.ZERO);
-  }
-
-  @Override
-  public BigDecimal calcularIvaVenta(
-      BusquedaFacturaVentaCriteria criteria, long idUsuarioLoggedIn) {
-    TipoDeComprobante[] tipoFactura = {TipoDeComprobante.FACTURA_A, TipoDeComprobante.FACTURA_B};
-    BigDecimal ivaVenta =
-        facturaVentaRepository.calcularIVAVenta(
-            this.getBuilderVenta(criteria, idUsuarioLoggedIn), tipoFactura);
-    return (ivaVenta != null ? ivaVenta : BigDecimal.ZERO);
-  }
-
-  @Override
-  public BigDecimal calcularIvaCompra(BusquedaFacturaCompraCriteria criteria) {
-    TipoDeComprobante[] tipoFactura = {TipoDeComprobante.FACTURA_A};
-    BigDecimal ivaCompra =
-        facturaCompraRepository.calcularIVACompra(this.getBuilderCompra(criteria), tipoFactura);
-    return (ivaCompra != null ? ivaCompra : BigDecimal.ZERO);
-  }
-
-  @Override
-  public BigDecimal calcularGananciaTotal(
-      BusquedaFacturaVentaCriteria criteria, long idUsuarioLoggedIn) {
-    BigDecimal gananciaTotal =
-        facturaVentaRepository.calcularGananciaTotal(
-            this.getBuilderVenta(criteria, idUsuarioLoggedIn));
-    return (gananciaTotal != null ? gananciaTotal : BigDecimal.ZERO);
-  }
-
-  @Override
   public BigDecimal calcularIVANetoRenglon(
       Movimiento movimiento,
       TipoDeComprobante tipo,
@@ -670,7 +272,7 @@ public class FacturaServiceImpl implements IFacturaService {
     } else if (movimiento == Movimiento.VENTA
         && (tipo == TipoDeComprobante.FACTURA_A
             || tipo == TipoDeComprobante.FACTURA_B
-    //        || tipo == TipoDeComprobante.FACTURA_Y
+            //        || tipo == TipoDeComprobante.FACTURA_Y
             || tipo == TipoDeComprobante.PRESUPUESTO)) {
       resultado =
           producto
@@ -680,8 +282,8 @@ public class FacturaServiceImpl implements IFacturaService {
                       .subtract(bonificacionPorcentaje.divide(CIEN, 15, RoundingMode.HALF_UP))
                       .multiply(
                           producto.getIvaPorcentaje().divide(CIEN, 15, RoundingMode.HALF_UP)));
-    //  if (tipo == TipoDeComprobante.FACTURA_Y)
-    //    resultado = resultado.divide(new BigDecimal("2"), 15, RoundingMode.HALF_UP);
+      //  if (tipo == TipoDeComprobante.FACTURA_Y)
+      //    resultado = resultado.divide(new BigDecimal("2"), 15, RoundingMode.HALF_UP);
     }
     return resultado;
   }
@@ -702,7 +304,7 @@ public class FacturaServiceImpl implements IFacturaService {
         if (tipo == TipoDeComprobante.FACTURA_A
             || tipo == TipoDeComprobante.FACTURA_B
             || tipo == TipoDeComprobante.FACTURA_C
-        //    || tipo == TipoDeComprobante.FACTURA_Y
+            //    || tipo == TipoDeComprobante.FACTURA_Y
             || tipo == TipoDeComprobante.PRESUPUESTO) {
           resultado =
               resultado.add(
@@ -746,7 +348,7 @@ public class FacturaServiceImpl implements IFacturaService {
         case FACTURA_X:
           resultado = producto.getPrecioVentaPublico();
           break;
-      /*  case FACTURA_Y:
+          /*  case FACTURA_Y:
           ivaResultado =
               producto
                   .getIvaPorcentaje()
@@ -764,139 +366,6 @@ public class FacturaServiceImpl implements IFacturaService {
       resultado = producto.getPrecioLista();
     }
     return resultado;
-  }
-
-  @Override
-  public long calcularNumeroFacturaVenta(
-      TipoDeComprobante tipoDeComprobante, long serie, long idSucursal) {
-    Long numeroFactura =
-        facturaVentaRepository.buscarMayorNumFacturaSegunTipo(tipoDeComprobante, serie, idSucursal);
-    if (numeroFactura == null) {
-      return 1; // No existe ninguna Factura anterior
-    } else {
-      return 1 + numeroFactura;
-    }
-  }
-
-  @Override
-  public byte[] getReporteFacturaVenta(Factura factura) {
-    ClassLoader classLoader = FacturaServiceImpl.class.getClassLoader();
-    InputStream isFileReport =
-        classLoader.getResourceAsStream("sic/vista/reportes/FacturaVenta.jasper");
-    Map<String, Object> params = new HashMap<>();
-    ConfiguracionSucursal configuracionSucursal =
-        this.configuracionSucursalService.getConfiguracionSucursal(factura.getSucursal());
-    params.put("preImpresa", configuracionSucursal.isUsarFacturaVentaPreImpresa());
-    if (factura.getTipoComprobante().equals(TipoDeComprobante.FACTURA_B)
-        || factura.getTipoComprobante().equals(TipoDeComprobante.PRESUPUESTO)) {
-      factura.setSubTotalBruto(factura.getTotal());
-      factura.setIva105Neto(BigDecimal.ZERO);
-      factura.setIva21Neto(BigDecimal.ZERO);
-    }
-    params.put("facturaVenta", factura);
-    if (factura.getTipoComprobante().equals(TipoDeComprobante.FACTURA_A)
-        || factura.getTipoComprobante().equals(TipoDeComprobante.FACTURA_B)
-        || factura.getTipoComprobante().equals(TipoDeComprobante.FACTURA_C)) {
-      if (factura.getNumSerieAfip() != 0 && factura.getNumFacturaAfip() != 0) {
-        params.put("nroSerie", factura.getNumSerieAfip());
-        params.put("nroFactura", factura.getNumFacturaAfip());
-      } else {
-        params.put("nroSerie", null);
-        params.put("nroFactura", null);
-      }
-    } else {
-      params.put("nroSerie", factura.getNumSerie());
-      params.put("nroFactura", factura.getNumFactura());
-    }
-    if (factura.getSucursal().getLogo() != null && !factura.getSucursal().getLogo().isEmpty()) {
-      try {
-        params.put(
-            "logo",
-            new ImageIcon(ImageIO.read(new URL(factura.getSucursal().getLogo()))).getImage());
-      } catch (IOException ex) {
-        logger.error(ex.getMessage());
-        throw new ServiceException(
-            messageSource.getMessage("mensaje_sucursal_404_logo", null, Locale.getDefault()), ex);
-      }
-    }
-    List<RenglonFactura> renglones = this.getRenglonesDeLaFactura(factura.getIdFactura());
-    JRBeanCollectionDataSource ds = new JRBeanCollectionDataSource(renglones);
-    try {
-      return JasperExportManager.exportReportToPdf(
-          JasperFillManager.fillReport(isFileReport, params, ds));
-    } catch (JRException ex) {
-      logger.error(ex.getMessage());
-      throw new ServiceException(
-          messageSource.getMessage("mensaje_error_reporte", null, Locale.getDefault()), ex);
-    }
-  }
-
-  @Override
-  public List<RenglonFactura> getRenglonesPedidoParaFacturar(
-      long idPedido, TipoDeComprobante tipoDeComprobante) {
-    List<RenglonFactura> renglonesRestantes = new ArrayList<>();
-    List<RenglonPedido> renglonesPedido =
-        pedidoService.getRenglonesDelPedidoOrdenadorPorIdRenglon(idPedido);
-    Map<Long, RenglonFactura> renglonesDeFacturas =
-        pedidoService.getRenglonesFacturadosDelPedido(idPedido);
-    if (renglonesDeFacturas != null) {
-      renglonesPedido.forEach(
-          r -> {
-            if (renglonesDeFacturas.containsKey(r.getIdProductoItem())) {
-              if (r.getCantidad()
-                      .compareTo(renglonesDeFacturas.get(r.getIdProductoItem()).getCantidad())
-                  > 0) {
-                renglonesRestantes.add(
-                    this.calcularRenglon(
-                        tipoDeComprobante,
-                        Movimiento.VENTA,
-                        r.getCantidad()
-                            .subtract(renglonesDeFacturas.get(r.getIdProductoItem()).getCantidad()),
-                        r.getIdProductoItem(),
-                        null));
-              }
-            } else {
-              renglonesRestantes.add(
-                  this.calcularRenglon(
-                      tipoDeComprobante,
-                      Movimiento.VENTA,
-                      r.getCantidad(),
-                      r.getIdProductoItem(),
-                      null));
-            }
-          });
-    } else {
-      renglonesPedido.forEach(
-          r ->
-              renglonesRestantes.add(
-                  this.calcularRenglon(
-                      tipoDeComprobante,
-                      Movimiento.VENTA,
-                      r.getCantidad(),
-                      r.getIdProductoItem(),
-                      null)));
-    }
-    return renglonesRestantes;
-  }
-
-  @Override
-  public boolean pedidoTotalmenteFacturado(Pedido pedido) {
-    boolean facturado = false;
-    Map<Long, RenglonFactura> renglonesDeFacturas =
-        pedidoService.getRenglonesFacturadosDelPedido(pedido.getIdPedido());
-    if (!renglonesDeFacturas.isEmpty()) {
-      for (RenglonPedido r : pedido.getRenglones()) {
-        if (renglonesDeFacturas.containsKey(r.getIdProductoItem())) {
-          facturado =
-              (r.getCantidad()
-                      .compareTo(renglonesDeFacturas.get(r.getIdProductoItem()).getCantidad())
-                  < 1);
-        } else {
-          return false;
-        }
-      }
-    }
-    return facturado;
   }
 
   @Override
@@ -921,7 +390,8 @@ public class FacturaServiceImpl implements IFacturaService {
     return renglones;
   }
 
-  private RenglonFactura calcularRenglon(
+  @Override
+  public RenglonFactura calcularRenglon(
       TipoDeComprobante tipoDeComprobante,
       Movimiento movimiento,
       BigDecimal cantidad,
@@ -944,9 +414,10 @@ public class FacturaServiceImpl implements IFacturaService {
     if (movimiento.equals(Movimiento.VENTA) || movimiento.equals(Movimiento.PEDIDO)) {
       this.asignarBonificacion(nuevoRenglon, producto);
     } else {
-      nuevoRenglon.setBonificacionPorcentaje(bonificacion != null ?  bonificacion : BigDecimal.ZERO);
+      nuevoRenglon.setBonificacionPorcentaje(bonificacion != null ? bonificacion : BigDecimal.ZERO);
       nuevoRenglon.setBonificacionNeta(
-          CalculosComprobante.calcularProporcion(nuevoRenglon.getPrecioUnitario(), nuevoRenglon.getBonificacionPorcentaje()));
+          CalculosComprobante.calcularProporcion(
+              nuevoRenglon.getPrecioUnitario(), nuevoRenglon.getBonificacionPorcentaje()));
     }
     nuevoRenglon.setIvaPorcentaje(producto.getIvaPorcentaje());
     /*if (tipoDeComprobante.equals(TipoDeComprobante.FACTURA_Y)) {
@@ -959,8 +430,8 @@ public class FacturaServiceImpl implements IFacturaService {
     nuevoRenglon.setGananciaPorcentaje(producto.getGananciaPorcentaje());
     nuevoRenglon.setGananciaNeto(producto.getGananciaNeto());
     nuevoRenglon.setImporteAnterior(
-            CalculosComprobante.calcularImporte(
-                    nuevoRenglon.getCantidad(), producto.getPrecioLista(), BigDecimal.ZERO));
+        CalculosComprobante.calcularImporte(
+            nuevoRenglon.getCantidad(), producto.getPrecioLista(), BigDecimal.ZERO));
     nuevoRenglon.setImporte(
         CalculosComprobante.calcularImporte(
             cantidad, nuevoRenglon.getPrecioUnitario(), nuevoRenglon.getBonificacionNeta()));
@@ -984,7 +455,7 @@ public class FacturaServiceImpl implements IFacturaService {
     // IVA
     if (nuevosResultadosComprobante.getTipoDeComprobante() == TipoDeComprobante.FACTURA_A
         || nuevosResultadosComprobante.getTipoDeComprobante() == TipoDeComprobante.FACTURA_B
-        //|| nuevosResultadosComprobante.getTipoDeComprobante() == TipoDeComprobante.FACTURA_Y
+        // || nuevosResultadosComprobante.getTipoDeComprobante() == TipoDeComprobante.FACTURA_Y
         || nuevosResultadosComprobante.getTipoDeComprobante() == TipoDeComprobante.PRESUPUESTO) {
       resultados.setIva21Neto(
           this.calcularIvaNetoFactura(
@@ -1043,346 +514,6 @@ public class FacturaServiceImpl implements IFacturaService {
     } else {
       nuevoRenglon.setBonificacionPorcentaje(BigDecimal.ZERO);
       nuevoRenglon.setBonificacionNeta(BigDecimal.ZERO);
-    }
-  }
-
-  @Override
-  public List<FacturaVenta> dividirFactura(FacturaVenta facturaADividir, int[] indices) {
-    FacturaVenta facturaSinIVA = new FacturaVenta();
-    facturaSinIVA.setCliente(facturaADividir.getCliente());
-    facturaSinIVA.setClienteEmbedded(facturaADividir.getClienteEmbedded());
-    facturaSinIVA.setUsuario(facturaADividir.getUsuario());
-    facturaSinIVA.setPedido(facturaADividir.getPedido());
-    facturaSinIVA.setDescuentoPorcentaje(facturaADividir.getDescuentoPorcentaje());
-    facturaSinIVA.setRecargoPorcentaje(facturaADividir.getRecargoPorcentaje());
-    FacturaVenta facturaConIVA = new FacturaVenta();
-    facturaConIVA.setCliente(facturaADividir.getCliente());
-    facturaConIVA.setClienteEmbedded(facturaADividir.getClienteEmbedded());
-    facturaConIVA.setUsuario(facturaADividir.getUsuario());
-    facturaConIVA.setPedido(facturaADividir.getPedido());
-    facturaConIVA.setTipoComprobante(facturaADividir.getTipoComprobante());
-    facturaConIVA.setDescuentoPorcentaje(facturaADividir.getDescuentoPorcentaje());
-    facturaConIVA.setRecargoPorcentaje(facturaADividir.getRecargoPorcentaje());
-    List<FacturaVenta> facturas = new ArrayList<>();
-    this.agregarRenglonesAFacturaSinIVA(facturaSinIVA, indices, facturaADividir.getRenglones());
-    this.agregarRenglonesAFacturaConIVA(facturaConIVA, indices, facturaADividir.getRenglones());
-    if (!facturaSinIVA.getRenglones().isEmpty()) {
-      this.procesarFacturaSinIVA(facturaADividir, facturaSinIVA);
-      facturas.add(facturaSinIVA);
-    }
-    this.procesarFacturaConIVA(facturaADividir, facturaConIVA);
-    facturas.add(facturaConIVA);
-    return facturas;
-  }
-
-  private FacturaVenta procesarFacturaSinIVA(
-      FacturaVenta facturaADividir, FacturaVenta facturaSinIVA) {
-    int size = facturaSinIVA.getRenglones().size();
-    BigDecimal[] importes = new BigDecimal[size];
-    BigDecimal[] cantidades = new BigDecimal[size];
-    BigDecimal[] ivaPorcentajeRenglones = new BigDecimal[size];
-    BigDecimal[] ivaNetoRenglones = new BigDecimal[size];
-    int indice = 0;
-    List<RenglonFactura> listRenglonesSinIVA = new ArrayList<>(facturaSinIVA.getRenglones());
-    facturaSinIVA.setFecha(facturaADividir.getFecha());
-    facturaSinIVA.setTipoComprobante(TipoDeComprobante.FACTURA_X);
-    facturaSinIVA.setFechaVencimiento(facturaADividir.getFechaVencimiento());
-    facturaSinIVA.setTransportista(facturaADividir.getTransportista());
-    facturaSinIVA.setRenglones(listRenglonesSinIVA);
-    for (RenglonFactura renglon : facturaSinIVA.getRenglones()) {
-      importes[indice] = renglon.getImporte();
-      cantidades[indice] = renglon.getCantidad();
-      ivaPorcentajeRenglones[indice] = renglon.getIvaPorcentaje();
-      ivaNetoRenglones[indice] = renglon.getIvaNeto();
-      indice++;
-    }
-    facturaSinIVA.setSubTotal(CalculosComprobante.calcularSubTotal(importes));
-    facturaSinIVA.setDescuentoNeto(
-        CalculosComprobante.calcularProporcion(
-            facturaSinIVA.getSubTotal(), facturaSinIVA.getDescuentoPorcentaje()));
-    facturaSinIVA.setRecargoNeto(
-        CalculosComprobante.calcularProporcion(
-            facturaSinIVA.getSubTotal(), facturaSinIVA.getRecargoPorcentaje()));
-    facturaSinIVA.setIva105Neto(
-        this.calcularIvaNetoFactura(
-            facturaSinIVA.getTipoComprobante(),
-            cantidades,
-            ivaPorcentajeRenglones,
-            ivaNetoRenglones,
-            IVA_105,
-            facturaADividir.getDescuentoPorcentaje(),
-            facturaADividir.getRecargoPorcentaje()));
-    facturaSinIVA.setIva21Neto(
-        this.calcularIvaNetoFactura(
-            facturaSinIVA.getTipoComprobante(),
-            cantidades,
-            ivaPorcentajeRenglones,
-            ivaNetoRenglones,
-            IVA_21,
-            facturaADividir.getDescuentoPorcentaje(),
-            facturaADividir.getRecargoPorcentaje()));
-    facturaSinIVA.setSubTotalBruto(
-        CalculosComprobante.calcularSubTotalBruto(
-            (facturaSinIVA.getTipoComprobante() == TipoDeComprobante.FACTURA_B
-                || facturaSinIVA.getTipoComprobante() == TipoDeComprobante.PRESUPUESTO),
-            facturaSinIVA.getSubTotal(),
-            facturaSinIVA.getRecargoNeto(),
-            facturaSinIVA.getDescuentoNeto(),
-            facturaSinIVA.getIva105Neto(),
-            facturaSinIVA.getIva21Neto()));
-    facturaSinIVA.setTotal(
-        CalculosComprobante.calcularTotal(
-            facturaSinIVA.getSubTotalBruto(),
-            facturaSinIVA.getIva105Neto(),
-            facturaSinIVA.getIva21Neto()));
-    facturaSinIVA.setObservaciones(facturaADividir.getObservaciones());
-    facturaSinIVA.setSucursal(facturaADividir.getSucursal());
-    facturaSinIVA.setEliminada(facturaADividir.isEliminada());
-    return facturaSinIVA;
-  }
-
-  private FacturaVenta procesarFacturaConIVA(
-      FacturaVenta facturaADividir, FacturaVenta facturaConIVA) {
-    int size = facturaConIVA.getRenglones().size();
-    BigDecimal[] importes = new BigDecimal[size];
-    BigDecimal[] cantidades = new BigDecimal[size];
-    BigDecimal[] ivaPorcentajeRenglones = new BigDecimal[size];
-    BigDecimal[] ivaNetoRenglones = new BigDecimal[size];
-    int indice = 0;
-    List<RenglonFactura> listRenglonesConIVA = new ArrayList<>(facturaConIVA.getRenglones());
-    facturaConIVA.setFecha(facturaADividir.getFecha());
-    facturaConIVA.setTipoComprobante(facturaADividir.getTipoComprobante());
-    facturaConIVA.setFechaVencimiento(facturaADividir.getFechaVencimiento());
-    facturaConIVA.setTransportista(facturaADividir.getTransportista());
-    facturaConIVA.setRenglones(listRenglonesConIVA);
-    for (RenglonFactura renglon : facturaConIVA.getRenglones()) {
-      importes[indice] = renglon.getImporte();
-      cantidades[indice] = renglon.getCantidad();
-      ivaPorcentajeRenglones[indice] = renglon.getIvaPorcentaje();
-      ivaNetoRenglones[indice] = renglon.getIvaNeto();
-      indice++;
-    }
-    facturaConIVA.setSubTotal(CalculosComprobante.calcularSubTotal(importes));
-    facturaConIVA.setDescuentoNeto(
-        CalculosComprobante.calcularProporcion(
-            facturaConIVA.getSubTotal(), facturaConIVA.getDescuentoPorcentaje()));
-    facturaConIVA.setRecargoNeto(
-        CalculosComprobante.calcularProporcion(
-            facturaConIVA.getSubTotal(), facturaConIVA.getRecargoPorcentaje()));
-    facturaConIVA.setIva105Neto(
-        this.calcularIvaNetoFactura(
-            facturaConIVA.getTipoComprobante(),
-            cantidades,
-            ivaPorcentajeRenglones,
-            ivaNetoRenglones,
-            IVA_105,
-            facturaADividir.getDescuentoPorcentaje(),
-            facturaADividir.getRecargoPorcentaje()));
-    facturaConIVA.setIva21Neto(
-        this.calcularIvaNetoFactura(
-            facturaConIVA.getTipoComprobante(),
-            cantidades,
-            ivaPorcentajeRenglones,
-            ivaNetoRenglones,
-            IVA_21,
-            facturaADividir.getDescuentoPorcentaje(),
-            facturaADividir.getRecargoPorcentaje()));
-    facturaConIVA.setSubTotalBruto(
-        CalculosComprobante.calcularSubTotalBruto(
-            (facturaConIVA.getTipoComprobante() == TipoDeComprobante.FACTURA_B
-                || facturaConIVA.getTipoComprobante() == TipoDeComprobante.PRESUPUESTO),
-            facturaConIVA.getSubTotal(),
-            facturaConIVA.getRecargoNeto(),
-            facturaConIVA.getDescuentoNeto(),
-            facturaConIVA.getIva105Neto(),
-            facturaConIVA.getIva21Neto()));
-    facturaConIVA.setTotal(
-        CalculosComprobante.calcularTotal(
-            facturaConIVA.getSubTotalBruto(),
-            facturaConIVA.getIva105Neto(),
-            facturaConIVA.getIva21Neto()));
-    facturaConIVA.setObservaciones(facturaADividir.getObservaciones());
-    facturaConIVA.setSucursal(facturaADividir.getSucursal());
-    facturaConIVA.setEliminada(facturaADividir.isEliminada());
-    return facturaConIVA;
-  }
-
-  private FacturaVenta agregarRenglonesAFacturaSinIVA(
-      FacturaVenta facturaSinIVA, int[] indices, List<RenglonFactura> renglones) {
-    List<RenglonFactura> renglonesSinIVA = new ArrayList<>();
-    BigDecimal cantidadProductosRenglonFacturaSinIVA = BigDecimal.ZERO;
-    int renglonMarcado = 0;
-    int numeroDeRenglon = 0;
-    for (RenglonFactura renglon : renglones) {
-      if (numeroDeRenglon == indices[renglonMarcado]) {
-        BigDecimal cantidad = renglon.getCantidad();
-        if (cantidad.compareTo(BigDecimal.ONE) >= 0) {
-          if ((cantidad.remainder(BigDecimal.ONE).compareTo(BigDecimal.ZERO) != 0)
-              || cantidad.remainder(new BigDecimal("2")).compareTo(BigDecimal.ZERO) == 0) {
-            cantidadProductosRenglonFacturaSinIVA =
-                cantidad.divide(new BigDecimal("2"), 15, RoundingMode.HALF_UP);
-          } else if (cantidad.remainder(new BigDecimal(2)).compareTo(BigDecimal.ZERO) != 0) {
-            cantidadProductosRenglonFacturaSinIVA =
-                cantidad.subtract(
-                    cantidad
-                        .divide(new BigDecimal("2"), 15, RoundingMode.HALF_UP)
-                        .setScale(0, RoundingMode.CEILING));
-          }
-        } else {
-          cantidadProductosRenglonFacturaSinIVA = BigDecimal.ZERO;
-        }
-        RenglonFactura nuevoRenglonSinIVA =
-            this.calcularRenglon(
-                TipoDeComprobante.FACTURA_X,
-                Movimiento.VENTA,
-                cantidadProductosRenglonFacturaSinIVA,
-                renglon.getIdProductoItem(),
-                null);
-        if (nuevoRenglonSinIVA.getCantidad().compareTo(BigDecimal.ZERO) != 0) {
-          renglonesSinIVA.add(nuevoRenglonSinIVA);
-        }
-        numeroDeRenglon++;
-        renglonMarcado++;
-        if (renglonMarcado == indices.length) {
-          break;
-        }
-      } else {
-        numeroDeRenglon++;
-      }
-    }
-    facturaSinIVA.setRenglones(renglonesSinIVA);
-    return facturaSinIVA;
-  }
-
-  private FacturaVenta agregarRenglonesAFacturaConIVA(
-      FacturaVenta facturaConIVA, int[] indices, List<RenglonFactura> renglones) {
-    List<RenglonFactura> renglonesConIVA = new ArrayList<>();
-    BigDecimal cantidadProductosRenglonFacturaConIVA = BigDecimal.ZERO;
-    int renglonMarcado = 0;
-    int numeroDeRenglon = 0;
-    for (RenglonFactura renglon : renglones) {
-      if (renglonMarcado < indices.length) {
-        if (numeroDeRenglon == indices[renglonMarcado]) {
-          BigDecimal cantidad = renglon.getCantidad();
-          if (cantidad.compareTo(BigDecimal.ONE) < 0 || cantidad.compareTo(BigDecimal.ONE) == 0) {
-            cantidadProductosRenglonFacturaConIVA = cantidad;
-          } else if ((cantidad.remainder(BigDecimal.ONE).compareTo(BigDecimal.ZERO) != 0)
-              || renglon.getCantidad().remainder(new BigDecimal(2)).compareTo(BigDecimal.ZERO)
-                  == 0) {
-            cantidadProductosRenglonFacturaConIVA =
-                renglon.getCantidad().divide(new BigDecimal("2"), 15, RoundingMode.HALF_UP);
-          } else if (renglon.getCantidad().remainder(new BigDecimal("2")).compareTo(BigDecimal.ZERO)
-              != 0) {
-            cantidadProductosRenglonFacturaConIVA =
-                renglon
-                    .getCantidad()
-                    .divide(new BigDecimal("2"), 15, RoundingMode.HALF_UP)
-                    .setScale(0, RoundingMode.CEILING);
-          }
-          renglonesConIVA.add(
-              calcularRenglon(
-                  facturaConIVA.getTipoComprobante(),
-                  Movimiento.VENTA,
-                  cantidadProductosRenglonFacturaConIVA,
-                  renglon.getIdProductoItem(),
-                  null));
-          renglonMarcado++;
-          numeroDeRenglon++;
-        } else {
-          numeroDeRenglon++;
-          renglonesConIVA.add(
-              calcularRenglon(
-                  facturaConIVA.getTipoComprobante(),
-                  Movimiento.VENTA,
-                  renglon.getCantidad(),
-                  renglon.getIdProductoItem(),
-                  null));
-        }
-      } else {
-        numeroDeRenglon++;
-        renglonesConIVA.add(
-            calcularRenglon(
-                facturaConIVA.getTipoComprobante(),
-                Movimiento.VENTA,
-                renglon.getCantidad(),
-                renglon.getIdProductoItem(),
-                null));
-      }
-    }
-    facturaConIVA.setRenglones(renglonesConIVA);
-    return facturaConIVA;
-  }
-
-  @Override
-  public boolean existeFacturaVentaAnteriorSinAutorizar(ComprobanteAFIP comprobante) {
-    QFacturaVenta qFacturaVenta = QFacturaVenta.facturaVenta;
-    BooleanBuilder builder = new BooleanBuilder();
-    builder.and(
-        qFacturaVenta
-            .idFactura
-            .lt(comprobante.getIdComprobante())
-            .and(qFacturaVenta.eliminada.eq(false))
-            .and(qFacturaVenta.sucursal.idSucursal.eq(comprobante.getSucursal().getIdSucursal()))
-            .and(qFacturaVenta.tipoComprobante.eq(comprobante.getTipoComprobante())));
-    Page<FacturaVenta> facturaAnterior =
-        facturaVentaRepository.findAll(
-            builder, PageRequest.of(0, 1, new Sort(Sort.Direction.DESC, "fecha")));
-    return facturaAnterior.getContent().get(0).getCae() == 0L;
-  }
-
-  @Override
-  public void enviarFacturaVentaPorEmail(long idFactura) {
-    Factura factura = this.getFacturaNoEliminadaPorId(idFactura);
-    List<TipoDeComprobante> tiposPermitidosParaEnviar =
-        Arrays.asList(
-            TipoDeComprobante.FACTURA_A, TipoDeComprobante.FACTURA_B, TipoDeComprobante.FACTURA_C);
-    if (factura instanceof FacturaVenta) {
-      if (tiposPermitidosParaEnviar.contains(factura.getTipoComprobante())
-          && factura.getCae() == 0L) {
-        throw new BusinessServiceException(
-            messageSource.getMessage("mensaje_correo_factura_sin_cae", null, Locale.getDefault()));
-      }
-      FacturaVenta facturaVenta = (FacturaVenta) factura;
-      if (facturaVenta.getCliente().getEmail() == null
-          || facturaVenta.getCliente().getEmail().isEmpty()) {
-        throw new BusinessServiceException(
-            messageSource.getMessage(
-                "mensaje_correo_cliente_sin_email", null, Locale.getDefault()));
-      }
-      String bodyEmail = "";
-      if (facturaVenta.getPedido() != null) {
-        if (facturaVenta.getPedido().getTipoDeEnvio().equals(TipoDeEnvio.RETIRO_EN_SUCURSAL)) {
-          bodyEmail =
-              messageSource.getMessage(
-                  "mensaje_correo_factura_retiro_sucursal",
-                  new Object[] {
-                    facturaVenta.getPedido().getSucursal().getNombre(),
-                    "(" + facturaVenta.getPedido().getDetalleEnvio().toString() + ")"
-                  },
-                  Locale.getDefault());
-        } else {
-          bodyEmail =
-              messageSource.getMessage(
-                  "mensaje_correo_factura_direccion_envio",
-                  new Object[] {facturaVenta.getPedido().getDetalleEnvio().toString()},
-                  Locale.getDefault());
-        }
-      } else {
-        bodyEmail =
-            messageSource.getMessage(
-                "mensaje_correo_factura_sin_pedido", null, Locale.getDefault());
-      }
-      correoElectronicoService.enviarEmail(
-          facturaVenta.getCliente().getEmail(),
-          "",
-          "Su Factura de Compra",
-          bodyEmail,
-          this.getReporteFacturaVenta(factura),
-          "Reporte");
-      logger.warn(
-          "El mail de la factura serie {} nro {} se envió.",
-          factura.getNumSerie(),
-          factura.getNumFactura());
     }
   }
 }
