+++ src/test/java/sic/integration/AppIntegrationTest.java
@@ -30,8 +30,8 @@ import sic.model.FacturaVenta;
 import sic.model.Pedido;
 import sic.model.RenglonPedido;
 import sic.modelo.*;
-import sic.modelo.calculos.NuevosResultadosComprobanteDTO;
-import sic.modelo.calculos.Resultados;
+import sic.modelo.dto.NuevosResultadosComprobanteDTO;
+import sic.modelo.Resultados;
 import sic.modelo.criteria.*;
 import sic.modelo.dto.*;
 import sic.service.ICajaService;
